// client/src/app/components/home/predators/predator-form/predator-form.model.ts
export class PredatorFormModel {
  predator?: number;
  total_number?: number;
  including_young_animals?: number;
  including_female_animals?: number;
  including_male_animals?: number;
  type_of_water?: number;
  distance_to_water?: number;
  territory?: number;
  activity?: number;
  comment?: string;
  gallery?: any;
}